+++ lib/glimmer-dsl-web/samples/hello/hello_component.rb
@@ -158,14 +158,12 @@ class AddressForm
         }
         
         style {
-          <<~CSS
-            #{address_div.selector} * {
-              margin: 5px;
-            }
-            #{address_div.selector} input, #{address_div.selector} select {
-              grid-column: 2;
-            }
-          CSS
+          r("#{address_div.selector} *") {
+            margin '5px'
+          }
+          r("#{address_div.selector} input, #{address_div.selector} select") {
+            grid_column '2'
+          }
         }
       }