+++ lib/glimmer-dsl-web/samples/hello/hello_content_data_binding.rb
@@ -117,20 +117,18 @@ Document.ready? do
               }
               
               style {
-                <<~CSS
-                  #{address_div.selector} {
-                    margin: 10px 0;
-                  }
-                  #{address_div.selector} * {
-                    margin: 5px;
-                  }
-                  #{address_div.selector} label {
-                    grid-column: 1;
-                  }
-                  #{address_div.selector} input, #{address_div.selector} select {
-                    grid-column: 2;
-                  }
-                CSS
+                r(address_div.selector) {
+                  margin '10px 0'
+                }
+                r("#{address_div.selector} *") {
+                  margin '5px'
+                }
+                r("#{address_div.selector} label") {
+                  grid_column '1'
+                }
+                r("#{address_div.selector} input, #{address_div.selector} select") {
+                  grid_column '2'
+                }
               }
             }
           }